Moved AsyncFunction to own class.
